requirements.txt
langchain>=0.0.252
chromadb>=0.4.3
openai>=0.27.8
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.4.2
pydantic>=2.0.3
requests>=2.31.0
numpy>=1.24.3
